NUEVO

+++ database/PROCEDIMIENTOS_TRIAJE.sql
@@ -40,11 +40,11 @@ CREATE PROCEDURE spu_triaje_agregar_triaje
 IN _idatencion 	INT,
 IN _idhistoria 	INT,
 IN _peso 	DECIMAL(5,2),
-IN _talla	DECIMAL(4,2),
+IN _talla	TINYINT,
 IN _frecuenciaCardiaca 		VARCHAR(5),
 IN _frecuenciaRespiratoria 	VARCHAR(5),
 IN _presionArterial 		VARCHAR(10),
-IN _temperatura			DECIMAL(4,2),
+IN _temperatura			DECIMAL(4,1),
 IN _saturacionOxigeno		TINYINT,
 IN _idusuario INT
 )
